feat(ActivitySection): move chart logic to component
